Use float16 for input array in memories_test.py.
A recent change (cl/969917669) has added support for
arbitrary/non-divisible index counts for unpacked sublane gathers,
so test test_sparsecore_unsupported_gather does not raise an exception
anymore.

This cupdates the test input in memories_test.py to use np.float16 instead of the default float type, to retain the UNIMPLEMENTED exception.

PiperOrigin-RevId: 970419441

+++ tests/memories_test.py
@@ -1891,7 +1891,7 @@ class SparsecoreOffloadTest(jtu.JaxTestCase):
       return jax.lax.gather(operand, indices, dnums, slice_sizes)
 
     inputs = (
-        np.linspace(0, 1, 10 * 5).reshape(10, 5),
+        np.linspace(0, 1, 10 * 5, dtype=np.float16).reshape(10, 5),
         np.array([[4, 2], [3, 2]]),
     )
